# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API doc of UniqueStudio Recruitment. For more API information, please see https://app.apifox.com/project/2985744",
        "title": "UniqueStudio Recruitment API",
        "contact": {
            "email": "wwbstar07@gmail.com"
        },
        "version": "0.1"
    },
    "host": "https://dev.back.recruitment2023.hustunique.com/",
    "paths": {
...
        "/recruitments/{rid}/interviews/{name}": {
            "get": {
                "description": "get recruitment interviews, candidate can't see slotNumber and selectNumber of Interviews (will get interviews of groups or unique), guarantee slotNumber \u003e selectNumber",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "get recruitment interviews.",
                "operationId": "get_recruitment_interviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recruitment id",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "web",
                            "lab",
                            "ai",
                            "game",
                            "mobile",
                            "android",
                            "ios",
                            "design",
                            "pm",
                            "unique"
                        ],
                        "type": "string",
                        "description": "pkg.Group",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/pkg.Interview"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            },
            "put": {
                "description": "get recruitment interviews, use PUt method to prevent resource are duplicated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "set recruitment interviews.",
                "operationId": "set_recruitment_interviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recruitment id",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "web",
                            "lab",
                            "ai",
                            "game",
                            "mobile",
                            "android",
                            "ios",
                            "design",
                            "pm",
                            "unique"
                        ],
                        "type": "string",
                        "description": "pkg.Group or unique",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update interview info",
                        "name": "[]pkg.UpdateInterviewOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.UpdateInterviewOpts"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            },
            "post": {
                "description": "create recruitment interviews.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "create recruitment interviews.",
                "operationId": "create_recruitment_interviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recruitment id",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "web",
                            "lab",
                            "ai",
                            "game",
                            "mobile",
                            "android",
                            "ios",
                            "design",
                            "pm",
                            "unique"
                        ],
                        "type": "string",
                        "description": "pkg.Group",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "create interview opts",
                        "name": "[]pkg.CreateInterviewOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.CreateInterviewOpts"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete recruitment interviews.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "delete recruitment interviews.",
                "operationId": "delete_recruitment_interviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recruitment id",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "web",
                            "lab",
                            "ai",
                            "game",
                            "mobile",
                            "android",
                            "ios",
                            "design",
                            "pm",
                            "unique"
                        ],
                        "type": "string",
                        "description": "pkg.Group or unique",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "delete interview opts",
                        "name": "[]pkg.DeleteInterviewOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.DeleteInterviewOpts"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "pkg.Application": {
            "type": "object",
            "properties": {
                "abandoned": {
                    "type": "boolean"
                },
                "answer": {
                    "type": "string"
                },
                "candidate_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "comments": {
                    "description": "onetomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Comment"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "grade": {
                    "description": "pkg.Grade",
                    "type": "string"
                },
                "group": {
                    "description": "pkg.Group",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.Group"
                        }
                    ]
                },
                "institute": {
                    "type": "string"
                },
                "interview_allocations_group": {
                    "$ref": "#/definitions/pkg.Interview"
                },
                "interview_allocations_group_id": {
                    "type": "string"
                },
                "interview_allocations_team": {
                    "$ref": "#/definitions/pkg.Interview"
                },
                "interview_allocations_team_id": {
                    "type": "string"
                },
                "interview_selections": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Interview"
                    }
                },
                "intro": {
                    "type": "string"
                },
                "is_quick": {
                    "type": "boolean"
                },
                "major": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "recruitment_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "referrer": {
                    "type": "string"
                },
                "rejected": {
                    "type": "boolean"
                },
                "resume": {
                    "type": "string"
                },
                "step": {
                    "description": "pkg.Step",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.Step"
                        }
                    ]
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_detail": {
                    "description": "get from sso",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.UserDetail"
                        }
                    ]
                }
            }
        },
        "pkg.Comment": {
            "type": "object",
            "properties": {
                "application_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "evaluation": {
                    "$ref": "#/definitions/pkg.Evaluation"
                },
                "member_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "member_name": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pkg.CreateInterviewOpts": {
            "type": "object",
            "required": [
                "date",
                "end",
                "period",
                "start"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/pkg.Period"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
        "pkg.DeleteInterviewOpts": {
            "type": "object",
            "required": [
                "iid"
            ],
            "properties": {
                "iid": {
                    "type": "string"
                }
            }
        },
        "pkg.Evaluation": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Good",
                "Normal",
                "Bad"
            ]
        },
        "pkg.Gender": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Male",
                "Female",
                "Oth"
            ]
        },
        "pkg.Group": {
            "type": "string",
            "enum": [
                "web",
                "lab",
                "ai",
                "game",
                "mobile",
                "android",
                "ios",
                "design",
                "pm",
                "unique"
            ],
            "x-enum-comments": {
                "Unique": "for team interview"
            },
            "x-enum-varnames": [
                "Web",
                "Lab",
                "Ai",
                "Game",
                "Mobile",
                "Android",
                "Ios",
                "Design",
                "Pm",
                "Unique"
            ]
        },
        "pkg.Interview": {
            "type": "object",
            "properties": {
                "applications": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Application"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/pkg.Group"
                },
                "period": {
                    "$ref": "#/definitions/pkg.Period"
                },
                "recruitment_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Period": {
            "type": "string",
            "enum": [
                "morning",
                "afternoon",
                "evening"
            ],
            "x-enum-varnames": [
                "Morning",
                "Afternoon",
                "Evening"
            ]
        },
...
        "pkg.Step": {
            "type": "string",
            "enum": [
                "SignUp",
                "WrittenTest",
                "GroupTimeSelection",
                "GroupInterview",
                "OnlineGroupInterview",
                "StressTest",
                "TeamTimeSelection",
                "TeamInterview",
                "OnlineTeamInterview",
                "Pass"
            ],
            "x-enum-comments": {
                "GroupInterview": "组面",
                "GroupTimeSelection": "组面时间选择",
                "OnlineGroupInterview": "在线组面",
                "OnlineTeamInterview": "在线群面",
                "Pass": "通过",
                "SignUp": "报名",
                "StressTest": "熬测",
                "TeamInterview": "群面",
                "TeamTimeSelection": "面试时间选择",
                "WrittenTest": "笔试"
            },
            "x-enum-varnames": [
                "SignUp",
                "WrittenTest",
                "GroupTimeSelection",
                "GroupInterview",
                "OnlineGroupInterview",
                "StressTest",
                "TeamTimeSelection",
                "TeamInterview",
                "OnlineTeamInterview",
                "Pass"
            ]
        },
...
        "pkg.UpdateInterviewOpts": {
            "type": "object",
            "required": [
                "date",
                "end",
                "period",
                "start"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/pkg.Period"
                },
                "start": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "pkg.UserDetail": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/pkg.Gender"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "join_time": {
                    "type": "string"
                },
                "lark_union_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
    },
...
}